fix(utils): splits functions into appropriate folders
