+++ client/components/utils/items.js
@@ -1,0 +1,102 @@
+export const gameItems = [
+  {
+    id: 1,
+    type: 'strawberry',
+    imageUrl: '/assets/strawberry.gif',
+    activeUrl: '/assets/strawberry.gif',
+    explodeUrl: '/assets/explodeRED.gif',
+    active: true,
+    x: 200,
+    y: 110,
+    width: 150
+  },
+  {
+    id: 2,
+    type: 'banana',
+    imageUrl: '/assets/banana.gif',
+    activeUrl: '/assets/banana.gif',
+    explodeUrl: '/assets/explodeYELLOW.gif',
+    active: true,
+    x: 1000,
+    y: 450,
+    width: 150
+  },
+  {
+    id: 3,
+    type: 'blackberry',
+    imageUrl: '/assets/blackberry.gif',
+    activeUrl: '/assets/blackberry.gif',
+    explodeUrl: '/assets/explodePURPLE.gif',
+    active: true,
+    x: 900,
+    y: 100,
+    width: 150
+  },
+  {
+    id: 4,
+    type: 'kiwi',
+    imageUrl: '/assets/kiwi.gif',
+    activeUrl: '/assets/kiwi.gif',
+    explodeUrl: '/assets/explodeGREEN.gif',
+    active: true,
+    x: 100,
+    y: 600,
+    width: 150
+  },
+  {
+    id: 5,
+    type: 'cherry',
+    imageUrl: '/assets/cherry.gif',
+    activeUrl: '/assets/cherry.gif',
+    explodeUrl: '/assets/explodeRED.gif',
+    active: true,
+    x: 1000,
+    y: 600,
+    width: 150
+  },
+  {
+    id: 6,
+    type: 'pineapple',
+    imageUrl: '/assets/pineapple.gif',
+    activeUrl: '/assets/pineapple.gif',
+    explodeUrl: '/assets/explodeYELLOW.gif',
+    active: true,
+    x: 950,
+    y: 200,
+    width: 150
+  },
+  {
+    id: 7,
+    type: 'apple',
+    imageUrl: '/assets/apple.gif',
+    activeUrl: '/assets/apple.gif',
+    explodeUrl: '/assets/explodeRED.gif',
+    active: true,
+    x: 110,
+    y: 150,
+    width: 150
+  },
+  {
+    id: 8,
+    type: 'watermelon',
+    imageUrl: '/assets/watermelon.gif',
+    activeUrl: '/assets/watermelon.gif',
+    explodeUrl: '/assets/explodeRED.gif',
+    active: true,
+    x: 1150,
+    y: 500,
+    width: 150
+  }
+]
+
+export const bomb = {
+  id: 9,
+  type: 'bomb',
+  imageUrl: '/assets/bomb.gif',
+  activeUrl: '/assets/bomb.gif',
+  explodeUrl: '/assets/boom.gif',
+  active: true,
+  x: 1050,
+  y: 20,
+  width: 150
+}
